remve write task in game
